@@ -6,5 +6,5 @@ use Symfony\Component\Validator\Constraint;
 
 class ValidCreditorReference extends Constraint
 {
-    public $message = 'The string "{{ string }}" is not a valid Creditor Reference.';
+    public string $message = 'The string "{{ string }}" is not a valid Creditor Reference.';
 }
