@@ -14,20 +14,20 @@ class AlternativeScheme implements QrCodeableInterface, SelfValidatableInterface
 
     /**
      * Parameter character chain of the alternative scheme
-     *
-     * @var string
      */
-    private $parameter;
+    private string $parameter;
+
+    private function __construct(string $parameter)
+    {
+        $this->parameter = $parameter;
+    }
 
     public static function create(string $parameter): self
     {
-        $alternativeScheme = new self();
-        $alternativeScheme->parameter = $parameter;
-
-        return $alternativeScheme;
+        return new self($parameter);
     }
 
-    public function getParameter(): ?string
+    public function getParameter(): string
     {
         return $this->parameter;
     }
